queryApp/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.17
django-filter==24.3
djangorestframework==3.15.2
dnspython==2.6.1
pymongo==4.10.1
sqlparse==0.5.3
typing_extensions==4.12.2
